requirements.txt
Flask
pandas
scikit-learn
requests
python-dotenv
numpy
matplotlib
seaborn